fix compilation on windoz
git-svn-id: http://svn.cern.ch/guest/lhcb/LHCb/trunk@22104 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 5203d4107fd760300d8be23a5096b5d3b572dd14

+++ LHCb/Muon/MuonInterfaces/MuonInterfaces/MuonLogHit.h
@@ -1,4 +1,4 @@
-// $Id: MuonLogHit.h,v 1.1 2010-02-10 19:07:50 ggiacomo Exp $
+// $Id: MuonLogHit.h,v 1.2 2010-02-11 09:29:51 ggiacomo Exp $
 #ifndef MUONLOGHIT_H 
 #define MUONLOGHIT_H 1
 #include "Kernel/MuonTileID.h"
@@ -28,7 +28,7 @@ class MuonLogHit {
   {m_ODEindex =(int)OdeIndex; }
   inline float time() const 
   { //avoid bias due to bins 0 and 1 being added in bin 1
-    return ((m_time+50*16)%16 == 1  ? m_time-0.499 : m_time);}
+    return (float) ((m_time+50*16)%16 == 1  ? m_time-0.499 : m_time);}
   inline int rawtime() const 
   { return m_time;}
   inline bool intime() const
